% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima.R
\name{all_disjoint_maxima}
\alias{all_disjoint_maxima}
\title{All sets of disjoint block maxima}
\usage{
all_disjoint_maxima(x, b = 1)
}
\arguments{
\item{x}{A numeric vector of raw observations.}

\item{b}{A numeric scalar.  The block size.}
}
\value{
A list containing
  \itemize{
    \item {\code{y} : } {the sliding block maxima.}
    \item {\code{x} : } {the subset of the input \code{x} that contributes
      to the values in \code{y}.}
  }
}
\description{
Calculates all sets of the maxima of disjoint blocks of \code{b} contiguous
values in the vector \code{x}.  In \code{\link{exdex}} this function
is used only for the purposes of checking \code{\link{all_maxima}}.
}
\details{
If \code{length(x)} is not an integer multiple of \code{b} then
  only the first \code{b * floor(n / b)} observations in \code{x} are used.
  \code{na.rm = TRUE} is passed to \code{max} so that blocks containing
  missing values produce a non-missing result.
}
\seealso{
\code{\link{spm}} for semiparametric estimation of the
  extremal index based on block maxima.

\code{\link{all_maxima}} for the calculation of both sliding and
  disjoint maxima.
}
