% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima.R
\name{all_maxima}
\alias{all_maxima}
\title{Sliding and disjoint block maxima}
\usage{
all_maxima(x, b = 1)
}
\arguments{
\item{x}{A numeric vector of raw observations.}

\item{b}{A numeric scalar.  The block size.}
}
\value{
A list containing
  \itemize{
    \item {\code{ys} : } {a numeric vector containing one set of sliding
      block maxima.}
    \item {\code{xs} : } {a numeric vector containing the values that
      contribute to \code{ys}, that is, the whole input vector \code{x}.}
    \item {\code{yd} : } {a \code{floor(n / b)} by
      \code{n - floor(n / b) * b + 1} numeric matrix.  Each column contains
      a set of disjoint maxima.}
    \item {\code{xd} : } {a \code{floor(n / b) * b} by
      \code{n - floor(n / b) * b + 1} numeric matrix.  Each column contains
      the values in \code{x} that contribute to the corresponding column
      in \code{ys}.}
  }
}
\description{
Calculates the (sliding) maxima of all blocks of \code{b} contiguous values
and all sets of the maxima of disjoint blocks of \code{b} contiguous values
in the vector \code{x}.  This provides the first step of computations in
\code{\link{spm}}.
}
\details{
\strong{Sliding maxima.} The function \code{\link[zoo]{rollapply}} in the
\code{zoo} package is used with argument \code{FUN = max} to calculate
  sliding (rolling) maxima.

  \strong{Disjoint maxima.}  If \code{n = length(x)} is an integer
  multiple of \code{b} then only one set of \code{n / b} disjoint
  block maxima are returned.
  Otherwise, \code{n - floor(n / b) * b + 1} sets of \code{floor(n / b)}
  disjoint block maxima are returned.  Set \code{i} are the disjoint maxima
  of \code{x[i:(i + floor(n / b) * b - 1)]}.  That is, all possible sets
  of contigous disjoint maxima achieving the maxima length of
  \code{floor(n / b)} are calculated.

  In both instances \code{na.rm = TRUE} is passed to \code{\link{max}} so
  that blocks containing missing values produce a non-missing result.

  Also returned are the values in \code{x} that contribute to each set
  of block maxima.
}
\seealso{
\code{\link{spm}} for semiparametric estimation of the
  extremal index based on block maxima.

\code{\link{sliding_maxima}} for the calculation of the maxima
  over sliding blocks.
}
