% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_b.R
\name{choose_b}
\alias{choose_b}
\title{Block length diagnostic plot}
\usage{
choose_b(data, b, bias_adjust = c("BB3", "BB1", "N", "none"),
  constrain = TRUE, varN = TRUE, level = 0.95,
  conf_scale = c("theta", "log"), interval_type = c("sym", "lik"),
  type = c("vertical", "cholesky", "spectral", "none"))
}
\description{
Description
}
\examples{
\dontrun{
b_vals <- seq(from = 50, to = 100, by = 5)
res <- choose_b(newlyn, b_vals)
plot(res)

b_vals <- seq(from = 25, to = 350, by = 25)
b_vals <- seq(from = 50, to = 350, by = 150)
res <- choose_b(sp500, b_vals)
plot(res)
}
}
