% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{coef.dgaps}
\alias{coef.dgaps}
\title{Extract Model Coefficients from a \code{"dgaps"} object}
\usage{
\method{coef}{dgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("kaps", "exdex")} returned from
\code{\link{dgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the estimate of the extremal index \eqn{\theta}.
}
\description{
\code{coef} method for class \code{c("dgaps", "exdex")}.
}
