% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwls_methods.R
\name{coef.iwls}
\alias{coef.iwls}
\title{Extract Model Coefficients from an \code{"iwls"} object}
\usage{
\method{coef}{iwls}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("kaps", "exdex")} returned from
\code{\link{iwls}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the estimate of the extremal index \eqn{\theta}.
}
\description{
\code{coef} method for class \code{c("iwls", "exdex")}.
}
