% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{coef.spm}
\alias{coef.spm}
\title{Extract Model Coefficients}
\usage{
\method{coef}{spm}(object, maxima = c("sliding", "disjoint"),
  estimator = c("N2015", "BB2018"), ...)
}
\arguments{
\item{object}{and object of class \code{c("spm", "exdex")} returned from
\code{\link{spm}}.}

\item{maxima}{A character scalar specifying whether to return the estimate
of the extremal index \eqn{\theta} based on sliding maxima or on disjoint
maxima.}

\item{estimator}{A character scalar specifying which variant of the
semiparametic maxima estimator to use: Northrop (2015) or
Berghaus and Bucher (2018).}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the required estimate of the extremal index
  \eqn{\theta}.
}
\description{
\code{coef} method for class \code{c("spm", "exdex")}.
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{https://doi.org/10.1007/s10687-015-0221-5}

Berghaus, B., Bucher, A. (2018) Weak convergence of a pseudo
maximum likelihood estimator for the extremal index. \emph{Ann. Statist.}
\strong{46}(5), 2307-2335. \url{https://doi.org/10.1214/17-AOS1621}
}
