% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confint.exdex}
\alias{confint.exdex}
\title{Confidence intervals for the extremal index \eqn{\theta}}
\usage{
\method{confint}{exdex}(object, level = 0.95, constrain = TRUE,
  conf_scale = c("theta", "log_theta"), ...)
}
\arguments{
\item{object}{An object of class \code{"exdex"}, returned by
\code{\link{spm}}.}

\item{level}{The confidence level required.  A numeric scalar in (0, 1).}

\item{constrain}{A logical scalar.  If \code{constrain = TRUE} then
any confidence limits that are greater than 1 are set to 1,
that is, they are constrained to lie in (0, 1].  Otherwise,
limits that are greater than 1 may be obtained.
If \code{constrain = TRUE} then any lower confidence limits that are
less than 0 are set to 0.}

\item{conf_scale}{A character scalar.   If \code{conf_scale = "theta"}
then confidence intervals are estimated for \eqn{\theta} directly.
If \code{conf_scale = "log_theta"} then confidence intervals are first
estimated for \eqn{log\theta} and then transformed back to the
\eqn{\theta}-scale.}

\item{...}{Further arguments.  None are used at present.}
}
\value{
A matrix with columns giving lower and upper confidence limits for
  each parameter. These will be labelled as (1 - level)/2 and
  1 - (1 - level)/2 in \% (by default 2.5\% and 97.5\%).
  The row names are the names of the model parameters,
  if these are available.
}
\description{
\code{confint} method for objects of class \code{"exdex"}.
Computes confidence intervals for \eqn{theta} based on an object returned
from \code{\link{spm}}.
}
\details{
Add details.
}
\examples{
res <- spm(newlyn, 20)
confint(res)
}
