% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima.R
\name{disjoint_maxima}
\alias{disjoint_maxima}
\title{Disjoint block maxima}
\usage{
disjoint_maxima(x, b = 1, which_dj = c("first", "last"))
}
\arguments{
\item{x}{A numeric vector of raw observations.}

\item{b}{A numeric scalar.  The block size.}

\item{which_dj}{A character scalar.  Determines Which set of disjoint
maxima are calculated: \code{"first"}, only the set whose first block
starts on the first observation in \code{x}; \code{"last"}, only the
set whose last block end on the last observation in \code{x}.}
}
\value{
A list containing
  \itemize{
    \item {\code{y} : } {the sliding block maxima.}
    \item {\code{x} : } {the subset of the input \code{x} that contributes
      to the values in \code{y}.}
  }
}
\description{
Calculates the maxima of disjoint blocks of \code{b} contiguous values in
the vector \code{x}.  In \code{\link{exdex}} this function
is used only for the purposes of checking \code{\link{all_maxima}}.
}
\details{
If \code{length(x)} is not an integer multiple of \code{b} then
  only the first \code{b * floor(n / b)} observations in \code{x} are used.
  \code{na.rm = TRUE} is passed to \code{max} so that blocks containing
  missing values produce a non-missing result.
}
\examples{
\dontrun{
x <- 1:11
disjoint_maxima(x, 3)
disjoint_maxima(x, 3, which_dj = "last")
}
}
\seealso{
\code{\link{spm}} for semiparametric estimation of the
  extremal index based on block maxima.

\code{\link{sliding_maxima}} for the calculation of the maxima
  over sliding blocks.
}
\keyword{internal}
