% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex-internal.R
\name{exdex-internal}
\alias{exdex-internal}
\alias{spm_check}
\alias{spm_sigmahat_dj}
\title{Internal exdex functions}
\usage{
spm_check(data, b, sliding = TRUE, bias_adjust = c("BB3", "BB1", "N",
  "none"), constrain = TRUE, varN = TRUE, which_dj = c("last",
  "first"))

spm_sigmahat_dj(data, b, dj_maxima, check = FALSE, which_dj)
}
\description{
These functions are not to be called by the user.
}
\keyword{internal}
