% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex.R
\docType{package}
\name{exdex}
\alias{exdex}
\alias{exdex-package}
\title{exdex: Estimation of the Extremal Index}
\description{
Performs frequentist inference about the extremal index using the
methodologies proposed in Northrop (2015), Suveges (2007) and
Suveges and Davison (2010).
}
\details{
Three
\itemize{
  \item{Semiparametric maxima estimator(Northrop, 2015)}
  \item{Iterated weighted least squares estimator: (Suveges, 2007)}
  \item{\eqn{K}-gaps estimator (Suveges and Davision, 2010)}
}

  See \code{vignette("exdex-vignette", package = "exdex")} for an
  overview of the package.
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{http://dx.doi.org/10.1007/s10687-015-0221-5}

Suveges, M. (2007) Likelihood estimation of the extremal
  index. \emph{Extremes}, \strong{10}, 41-55.
  \url{http://dx.doi.org/10.1007/s10687-007-0034-2}

Suveges, M. and Davison, A. C. (2010) Model
  misspecification in peaks over threshold analysis, \emph{The Annals of
  Applied Statistics}, \strong{4}(1), 203-221.
  \url{http://dx.doi.org/10.1214/09-AOAS292}
}
