% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decluster.R
\name{find_clusters}
\alias{find_clusters}
\title{Declustering}
\usage{
find_clusters(data, thresh, k = 1)
}
\arguments{
\item{data}{A numeric vector of raw data.  No missing values are allowed.}

\item{thresh}{A numeric scalar.  Extreme value threshold applied to data.}

\item{k}{An integer scalar}
}
\description{
Describe
}
\examples{
thresh <- quantile(newlyn, probs = 0.99)
#x <- find_clusters(newlyn, thresh)
}
