% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_matrix_test.R
\name{kgaps_imt}
\alias{kgaps_imt}
\title{Information matrix test under the K-gaps model}
\usage{
kgaps_imt(data, thresh, k = 1)
}
\arguments{
\item{data}{A numeric vector of raw data.  No missing values are allowed.}

\item{thresh, k}{Numeric vectors.  \code{thresh} is a vector of
extreme value thresholds applied to data.  \code{k} is a vector of values
of the run parameter \eqn{K}, as defined in Suveges and Davison (2010).
See \code{\link{kgaps_mle}} for more details.
The information matrix test is performed a over grid of all
combinations of threshold and \eqn{K} in the vectors \code{thresh}
and \code{k}.}
}
\value{
A list containing
}
\description{
Performs the information matrix test of Suveges and Davison (2010) to
diagnose misspecification of the K-gaps model
}
\details{
Add details
}
\examples{
thresh <- quantile(newlyn, probs = 0.90)
imt <- kgaps_imt(newlyn, thresh)
}
\references{
Suveges, M. and Davison, A. C. (2010) Model
  misspecification in peaks over threshold analysis, \emph{The Annals of
  Applied Statistics}, \strong{4}(1), 203-221.
  \url{https://doi.org/10.1214/09-AOAS292"}
}
\seealso{
\code{\link{kgaps_mle}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.
}
