% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{logLik.dgaps}
\alias{logLik.dgaps}
\title{Extract log-likelihood for objects of class \code{"dgaps"}}
\usage{
\method{logLik}{dgaps}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"dgaps"}, a result of a call to
\code{\link{dgaps}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
An object of class \code{"logLik"}: a numeric scalar with
value equal to the maximised log-likelihood.  The returned object also has
attributes \code{nobs}, the numbers of \eqn{K}-gaps that contribute to the
log-likelihood and \code{"df"}, which is equal to the number of total number
of parameters estimated (1).
}
\description{
\code{nobs} method for class \code{c("dgaps", "exdex")}.
}
