% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex.R
\docType{data}
\name{newlyn}
\alias{newlyn}
\title{Newlyn sea surges}
\format{A vector of length 2894.}
\source{
Coles, S.G. (1991) Modelling extreme multivariate events. PhD thesis,
  University of Sheffield, U.K.
}
\usage{
newlyn
}
\description{
The vector \code{newlyn} contains 2894 maximum sea-surges measured at
Newlyn, Cornwall, UK over the period 1971-1976. The observations are
the maximum hourly sea-surge heights over contiguous 15-hour time
periods.
}
\references{
Fawcett, L. and Walshaw, D. (2012) Estimating return levels from
  serially dependent extremes. \emph{Environmetrics}, \strong{23}(3),
  272-283.  \url{https://doi.org/10.1002/env.2133}

Northrop, P. J. (2015) An efficient semiparametric maxima
  estimator of the extremal index. \emph{Extremes}, \strong{18},
  585-603.  \url{https://doi.org/10.1007/s10687-015-0221-5}
}
\keyword{datasets}
