% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{nobs.dgaps}
\alias{nobs.dgaps}
\title{Extract the Number of Observations from a \code{"dgaps"} object}
\usage{
\method{nobs}{dgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("dgaps", "exdex")} returned from
\code{\link{dgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the number of inter-exceedance times used in the
  fit. If \code{x$inc_cens = TRUE} then this includes 2 censored
  observations.
}
\description{
\code{nobs} method for class \code{c("dgaps", "exdex")}.
}
