% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwls_methods.R
\name{nobs.iwls}
\alias{nobs.iwls}
\title{Extract the Number of Observations from an \code{"iwls"} object}
\usage{
\method{nobs}{iwls}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("iwls", "exdex")} returned from
\code{\link{iwls}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the number of inter-exceedance times used in the
  fit.
}
\description{
\code{nobs} method for class \code{c("iwls", "exdex")}.
}
