% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgaps_methods.R
\name{nobs.kgaps}
\alias{nobs.kgaps}
\title{Extract the Number of Observations from a \code{"kgaps"} object}
\usage{
\method{nobs}{kgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("kgaps", "exdex")} returned from
\code{\link{kgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the number of inter-exceedance times used in the
  fit. If \code{x$inc_cens = TRUE} then this includes 2 censored
  observations.
}
\description{
\code{nobs} method for class \code{c("kgaps", "exdex")}.
}
