% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_methods.R
\name{nobs.spm}
\alias{nobs.spm}
\title{Extract the Number of Observations from an \code{"spm"} object}
\usage{
\method{nobs}{spm}(object, maxima = c("sliding", "disjoint"), ...)
}
\arguments{
\item{object}{and object of class \code{c("spm", "exdex")} returned from
\code{\link{spm}}.}

\item{maxima}{A character scalar specifying whether to return the
number of observed sliding maxima or disjoint maxima.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the number of observations used in the fit.
}
\description{
\code{nobs} method for class \code{c("spm", "exdex")}.
}
