% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_uk.R
\name{plot.choose_uk}
\alias{plot.choose_uk}
\title{Plot block length diagnostic for the semiparametric maxima estimator}
\usage{
\method{plot}{choose_uk}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{c("choose_uk", "exdex")}, a result of a
call to \code{\link{choose_uk}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to
\code{\link[graphics]{matplot}} and/or \code{\link[graphics]{axis}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for objects inheriting from class \code{"choose_uk"},
returned from \code{\link{choose_uk}}
}
\details{
Add details
}
\section{Examples}{

See the examples in \code{\link{choose_uk}}.
}

\seealso{
\code{\link{choose_uk}}.
}
