% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{plot.confint_dgaps}
\alias{plot.confint_dgaps}
\title{Plot diagnostics for a confint_dgaps object}
\usage{
\method{plot}{confint_dgaps}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{c("confint_dgaps", "exdex")}, a result of
a call to \code{\link{confint.dgaps}}.}

\item{y}{Not used.}

\item{...}{Further arguments to be passed to
\code{\link[chandwich]{plot.confint}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an objects of class
\code{c("confint_dgaps", "exdex")}.
}
\section{Examples}{

See the examples in \code{\link{confint.dgaps}}.
}

\seealso{
\code{\link{confint.dgaps}}: \code{confint} method for
  class \code{c("dgaps", "exdex")}.
}
