% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{plot.confint_kgaps}
\alias{plot.confint_kgaps}
\title{Plot diagnostics for a confint_kgaps object}
\usage{
\method{plot}{confint_kgaps}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{c("confint_kgaps", "exdex")}, a result of
a call to \code{\link{confint.kgaps}}.}

\item{y}{Not used.}

\item{...}{Further arguments to be passed to
\code{\link[chandwich]{plot.confint}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an objects of class
\code{c("confint_kgaps", "exdex")}.
}
\section{Examples}{

See the examples in \code{\link{confint.kgaps}}.
}

\seealso{
\code{\link{confint.kgaps}}: \code{confint} method for
  class \code{c("kgaps", "exdex")}.
}
