% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_b.R
\name{plot.exdex}
\alias{plot.exdex}
\title{Plot diagnostics for an exdex object}
\usage{
\method{plot}{exdex}(x, y, ..., estimator = c("N2015", "BB2018"),
  maxima = c("sliding", "disjoint"))
}
\arguments{
\item{x}{an object of class "exdex", a result of a call to
\code{\link{choose_b}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to
\code{\link[graphics]{matplot}} and/or \code{\link[graphics]{axis}}.}

\item{estimator}{Choice of estimator: \code{"N2015"} for Northrop (2015),
\code{"BB2018"} for Berghaus and Bucher (2018).
See \code{\link{spm}} for details.}

\item{maxima}{Should the estimator be based on sliding or disjoint maxima?}
}
\value{
In addition to producing the plot a list of the arguments used
  by \code{\link[graphics]{matplot}}, \code{\link[graphics]{axis}} is
  returned (invisibly).
}
\description{
\code{plot} method for objects of class "exdex" returned from
\code{\link{choose_b}}
}
\details{
Produces a simple diagnostic plot to aid the choice of block
  length \eqn{b} based on the object returned from \code{\link{choose_b}}.
  Estimates of \eqn{b} and approximate \code{conf}\% confidence intervals
  are plotted against the value of \eqn{b} used to produce each estimate.
  The type of confidence interval is determined by the arguments
  \code{interval_type}, \code{conf_scale} and \code{type} provided in the
  call to \code{\link{choose_b}}.
}
\examples{
b_vals <- seq(from = 50, to = 350, by = 150)
res <- choose_b(sp500, b_vals)
plot(res)
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{https://doi.org/10.1007/s10687-015-0221-5}

Berghaus, B., Bucher, A. (2018) Weak convergence of a pseudo
maximum likelihood estimator for the extremal index. \emph{Ann. Statist.}
\strong{46}(5), 2307-2335. \url{https://doi.org/10.1214/17-AOS1621}
}
\seealso{
\code{\link{choose_b}}.
}
