% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_matrix_test.R
\name{plot.kgaps_imt}
\alias{plot.kgaps_imt}
\title{Plot method for object of class \code{"kgaps_imt"}}
\usage{
\method{plot}{kgaps_imt}(x, y = c("imt", "pvalue", "theta"),
  horiz_axis = c("k", "thresh"), alpha = 0.05, ...)
}
\arguments{
\item{x}{an object of class \code{c("kgaps_imt", "exdex")}, a result of a
call to \code{\link{kgaps_imt}}.}

\item{y}{A character scalar.   What should we plot on the vertical axis:
\code{"imt"} for the information matrix test statistic,
\code{"pvalue"} for the \eqn{p}-value associated with the information
  matrix test statistic,
\code{"theta"} for the K-gaps estimate of \eqn{theta}.}

\item{horiz_axis}{A character scalar.  Should we plot \eqn{K} or
threshold on the horizontal axis?}

\item{alpha}{A numeric scalar in (0, 1).  Significance level.}

\item{...}{Additional arguments.}
}
\value{
In addition to producing the plot a list of the arguments used
  by \code{\link[graphics]{matplot}}, \code{\link[graphics]{axis}} is
  returned (invisibly).
}
\description{
\code{plot} method for objects inheriting from class \code{"kgaps_imt"},
returned from \code{\link{kgaps_imt}}
}
\examples{
thresh <- quantile(newlyn, probs = seq(0.1, 0.9, by = 0.1))
imt <- kgaps_imt(newlyn, thresh, k = 1:5)
plot(imt)
}
\seealso{
\code{\link{kgaps_imt}}.
}
