% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{print.confint_dgaps}
\alias{print.confint_dgaps}
\title{Print method for a confint_dgaps object}
\usage{
\method{print}{confint_dgaps}(x, ...)
}
\arguments{
\item{x}{an object of class \code{c("confint_dgaps", "exdex")}, a result of
a call to \code{\link{confint.dgaps}}.}

\item{...}{Additional optional arguments to be passed to
\code{\link{print.default}}}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class \code{c("confint_dgaps", "exdex")}.
}
\details{
Prints the matrix of confidence intervals for \eqn{\theta}.
}
\seealso{
\code{\link{dgaps}} for estimation of the extremal index
  \eqn{\theta} using a semiparametric maxima method.

\code{\link{confint.dgaps}}: \code{confint} method for
  class \code{"dgaps"}.
}
