% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{print.confint_spm}
\alias{print.confint_spm}
\title{Print method for a confint_spm object}
\usage{
\method{print}{confint_spm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{c("confint_spm", "exdex")}, a result of
a call to \code{\link{adjust_loglik}}.}

\item{...}{Additional optional arguments to be passed to
\code{\link{print}}.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class \code{c("confint_spm", "exdex")}.
}
\details{
Just prints the original call to \code{\link{adjust_loglik}}
  and a character vector giving the names of the attributes
  (produced using \code{ls(attributes(x))}) to the function returned
  from \code{\link{adjust_loglik}}.
  To view an individual attribute called \code{att_name} use
  \code{attr(x, "att_name")} or \code{attributes(x)$att_name}.
}
\seealso{
\code{\link{spm}} for estimation of the extremal index
  \eqn{\theta} using a semiparametric maxima method.

\code{\link{confint.spm}}: \code{confint} method for
  class \code{"spm"}.
}
