% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{print.dgaps}
\alias{print.dgaps}
\title{Print method for a \code{"dgaps"} object}
\usage{
\method{print}{dgaps}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{c("dgaps", "exdex")}, a result of
a call to \code{\link{dgaps}}.}

\item{digits}{The argument \code{digits} to \code{\link{print.default}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class \code{c("dgaps", "exdex")}.
}
\details{
Prints the original call to \code{\link{dgaps}}
  and the estimate of the extremal index \eqn{\theta}.
}
\seealso{
\code{\link{dgaps}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the \eqn{K}-gaps model.

\code{\link{confint.dgaps}}: \code{confint} method for
  class \code{"dgaps"}.
}
