% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{print.summary.dgaps}
\alias{print.summary.dgaps}
\title{Print method for objects of class \code{"summary.dgaps"}}
\usage{
\method{print}{summary.dgaps}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.pm", a result of a call to
\code{\link{summary.dgaps}}.}

\item{...}{Additional arguments passed on to \code{\link{print.default}}.}
}
\value{
Prints the numeric matrix \code{x$summary} returned from
\code{\link{summary.dgaps}}.
}
\description{
\code{print} method for an object \code{x} of class \code{"summary.dgaps"}.
}
\section{Examples}{

See the examples in \code{\link{dgaps}}.
}

\seealso{
\code{\link{dgaps}} for estimation of the extremal index
  \eqn{\theta} using a semiparametric maxima method.

\code{\link{confint.dgaps}} for estimation of confidence intervals
  for \eqn{\theta}.
}
