% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgaps_methods.R
\name{print.summary.kgaps}
\alias{print.summary.kgaps}
\title{Print method for objects of class \code{"summary.kgaps"}}
\usage{
\method{print}{summary.kgaps}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.pm", a result of a call to
\code{\link{summary.kgaps}}.}

\item{...}{Additional arguments passed on to \code{\link{print.default}}.}
}
\value{
Prints the numeric matrix \code{x$summary} returned from
\code{\link{summary.kgaps}}.
}
\description{
\code{print} method for an object \code{x} of class \code{"summary.kgaps"}.
}
\section{Examples}{

See the examples in \code{\link{kgaps}}.
}

\seealso{
\code{\link{kgaps}} for estimation of the extremal index
  \eqn{\theta} using a semiparametric maxima method.

\code{\link{confint.kgaps}} for estimation of confidence intervals
  for \eqn{\theta}.
}
