% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima.R
\name{sliding_maxima}
\alias{sliding_maxima}
\title{Sliding block maxima}
\usage{
sliding_maxima(x, b = 1)
}
\arguments{
\item{x}{A numeric vector of raw observations.}

\item{b}{A numeric scalar.  The block size.}
}
\value{
A list containing
  \itemize{
    \item {\code{y} : } {a numeric vector of the sliding block maxima.}
    \item {\code{x} : } {the input \code{x}.}
  }
}
\description{
Calculates the maxima of all blocks of \code{b} contiguous values in the
vector \code{x}. In \code{\link{exdex}} this function is used within
\code{\link{all_maxima}} to provide the first step of computations in
\code{\link{spm}}.
}
\details{
The function \code{\link[zoo]{rollapply}} in the
\code{zoo} package is used with argument \code{FUN = max} to calculate
  sliding (rolling) maxima.  \code{na.rm = TRUE} is passed to
  \code{max} so that blocks containing missing values produce a
  non-missing result.
}
\examples{
\dontrun{
x <- 1:11
sliding_maxima(x, 3)
}
}
\references{
Zeileis, A. and Grothendieck, G. (2005). zoo: S3
  Infrastructure for Regular and Irregular Time Series. \emph{Journal of
  Statistical Software}, \strong{14}(6), 1-27.
  \url{https://doi.org/10.18637/jss.v014.i06}
}
\seealso{
\code{\link{spm}} for semiparametric estimation of the
  extremal index based on block maxima.

\code{\link{all_maxima}} for the calculation of both sliding and
  disjoint maxima.
}
\keyword{internal}
