% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Daily log returns of the Standard and Poor (S&P) 500 index}
\format{A vector of length 7250, created using \code{\link[zoo]{zoo}}
  with an "index" attribute giving the date of the corresponding negated
  log return.}
\source{
Yahoo finance \url{https://finance.yahoo.com/quote/^SPX/history/}
}
\usage{
sp500
}
\description{
Daily log returns of the S&P 500 index, that is, the log of the ratio of
successive daily closing prices, from 3rd January 1990 to 9th October 2018.
}
\keyword{datasets}
