% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_maxima.R
\name{spm}
\alias{spm}
\title{Semiparametric maxima estimator of the extremal index}
\usage{
spm(data, b, sliding = TRUE, bias_adjust = c("BB3", "BB1", "N",
  "none"), constrain = TRUE, conf = 0.95, conf_scale = c("theta",
  "log_theta"))
}
\arguments{
\item{data}{A numeric vector of raw data.}

\item{b}{A numeric scalar.  The block size.}

\item{sliding}{A logical scalar.  If \code{sliding = TRUE} then the
estimates are based on sliding block maxima.  Otherwise, the estimates
are based on disjoint block maxima.}

\item{bias_adjust}{A character scalar.  Is bias-adjustment of the
raw estimate of \eqn{\theta} performed using the bias-reduced
estimator (\code{bias_adjust = "BB3"}), derived in Section 5 of
Berghaus and Bucher (2018); or a simpler version of this, in which
the raw estimate is multiplied by \eqn{(k-1) / k}, where \eqn{k} is the
number of blocks (\code{bias_adjust = "BB1"}); or the bias-adjustment
of the empirical distribution function used to calculate the estimate,
as detailed in Section 2 of Northrop (2015).  When disjoint maxima are
used \code{bias_adjust = "BB1"} and \code{bias_adjust = "N"}
give identical estimates of the Berghaus and Bucher (2018) variant,
as explained at the end of Section 5 of Berghaus and Bucher (2018).
If \code{bias_adjust = "none"} then no bias-adjustment is performed.}

\item{constrain}{A logical scalar.  If \code{constrain = TRUE} then
any estimates or confidence limits that are greater than 1 are set to 1,
that is, they are constrained to lie in (0, 1].  This is carried out
\emph{after} any bias-adjustment.  Otherwise,
estimates and confidence limits that are greater than 1 may be obtained.
If \code{constrain = TRUE} then any lower confidence limits that are
less than 0 are set to 0.}

\item{conf}{A numeric scalar.  The confidence level required:
100\code{conf}\% confidence intervals are calculated.}

\item{conf_scale}{A character scalar.   If \code{conf_scale = "theta"}
then confidence intervals are estimated for \eqn{\theta} directly.
If \code{conf_scale = "log_theta"} then confidence intervals are first
estimated for \eqn{log\theta} and then transformed back to the
\eqn{\theta}-scale.}
}
\value{
A list containing
  \itemize{
    \item {\code{theta_mle} : } {The maximum likelihood estimate (MLE) of
      \eqn{\theta}.}
    \item {\code{theta_se} : } {The estimated standard error of the MLE.}
    \item {\code{theta_ci} : } {(If \code{conf} is supplied) a numeric
      vector of length two giving lower and upper confidence limits for
      \eqn{\theta}.}
  }
  If \code{conf} is not supplied then only the MLE \code{theta_mle}
  is returned.
}
\description{
Estimates the extremal index \eqn{\theta} using a semiparametric block
maxima estimator of Northrop (2015) and a variant of this estimator
studied by Berghaus and Bucher (2018).  Estimates of uncertainty are
calculated using the asymptotic theory developed by Berghaus and
Bucher (2018) and (optionally) a block bootstrap.
}
\details{
The extremal index \eqn{\theta} is estimated using the
  semiparametric maxima estimator of Northrop (2015).
  If \code{sliding = TRUE} then the function uses sliding block maxima,
  that is, the largest value observed in \emph{all}
  (\code{length(data) - b + 1}) blocks of \code{b} observations.
  If \code{sliding = FALSE} then disjoint block maxima, that is, the
  largest values in (\code{floor(n / b)}) disjoint blocks of \code{b}
  observations, are used.

  Estimation of the sampling variances of the estimators is based
  on Proposition 4.1 of Berghaus and Bucher (2018).  A condition imposed
  in this proposition means that that \code{b} must be no smaller than
  \eqn{k^{1/2}} and no larger than \eqn{k^2}, where \eqn{k} is
  \code{floor(length(data) / b)}, i.e. \eqn{k} is the number of complete
  blocks.  If this is not the case then a warning will be given and
    \itemize{
      \item{estimates of uncertainty (estimated standard
             errors and confidence intervals) will be missing from the
             returned object,}
      \item{if \code{bias_adjust == "BB3"} then this bias-adjustment
            based on \code{bias_adjust == "BB1"} will be performed instead,
            because the former relies on the estimated variances of the
            estimators.}
    }
}
\examples{
spm(-as.vector(sp500[2:6550]), 250)

spm(newlyn, 20)
spm(newlyn, 20, sliding = FALSE)

spm(newlyn, 20, sliding = FALSE, conf = 95)
# When sliding = TRUE bootstrapping is slow
\dontrun{
spm(newlyn, 20, conf = 95)
}
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{http://dx.doi.org/10.1007/s10687-015-0221-5}

Berghaus, B., Bucher, A. (2018) Weak convergence of a pseudo
maximum likelihood estimator for the extremal index. \emph{Ann. Statist.}
\strong{46}(5), 2307-2335. \url{http://dx.doi.org/10.1214/17-AOS1621}
}
\seealso{
\code{\link{kgaps_mle}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.
}
