% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_maxima.R
\name{spm_mle}
\alias{spm_mle}
\title{Semiparametric maxima estimator of the extremal index}
\usage{
spm_mle(data, b, sliding = TRUE, constrain = TRUE)
}
\arguments{
\item{data}{A numeric vector of raw data.}

\item{b}{A numeric scalar.  The block size.}

\item{sliding}{A logical scalar.}

\item{constrain}{A logical scalar.}
}
\value{
A list containing
  \itemize{
    \item {\code{theta_mle} : } {The maximum likelihood estimate (MLE) of
      \eqn{\theta}.}
    \item {\code{theta_se} : } {The estimated standard error of the MLE.}
    \item {\code{theta_ci} : } {(If \code{conf} is supplied) a numeric
      vector of length two giving lower and upper confidence limits for
      \eqn{\theta}.}
  }
}
\description{
Calculates the semiparametric maxima estimator of the extremal index
\eqn{\theta} based on sliding or disjoint block maxima.
}
\details{
If \code{sliding = TRUE} then the function uses sliding block maxima
whereas if \code{sliding = FALSE} then disjoint block maxima are used
instead.

If \code{constrain = TRUE} then ...
}
\examples{
spm_mle(newlyn, 20)
spm_mle(newlyn, 20, sliding = FALSE)
}
\references{
Northrop, P. J. (2015) \emph{An efficient semiparametric maxima
estimator of the extremal index} Extremes, \strong{18}(4), 585-603.
\url{http://dx.doi.org/10.1007/s10687-015-0221-5}
}
\seealso{
\code{\link{kgaps_mle}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.
}
