% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_maxima.R
\name{spm_sigmahat_dj}
\alias{spm_sigmahat_dj}
\title{Estimates \eqn{\sigma_{dj}^2}}
\usage{
spm_sigmahat_dj(data, b, dj_maxima, check = FALSE)
}
\arguments{
\item{data}{A numeric vector of raw data.}

\item{b}{A numeric scalar.  The block size.}

\item{dj_maxima}{A list returned from a call
\code{disjoint_maxima(data, b)} to \code{\link{disjoint_maxima}}.
If this isn't supplied then this call is made inside
\code{spm_sigmahat_dj}.}

\item{check}{A logical scalar.  Used only to check this function during
development.  If \code{check = TRUE} then a matrix with 5 columns is
returned, each columns containing the result of 5 different ways to
calculate the middle term of Bhatnj in the second line of the equation
displayed midpage on page 2319 of Berghaus and Bucher (2018).}
}
\value{
A vector of length two.  The first component is the estimate
 given by the first part of the penultimate displayed equation on page
 2319 of Berghaus and Bucher (2018). The second component is an analogous
 estimate based on the estimator proposed in Northrop (2015).
}
\description{
Estimates the value of \eqn{\sigma_{dj}^2}, a variance involved in
Sections 3, 4 and 5 of Berghaus and Bucher (2018).
These estimates are required in \code{\link{spm}}, which estimates
the extremal index using the methods described in Northrop (2015) and
Berghaus and Bucher (2018), in order to perform bias-adjustment and
estimation of uncertainty.
}
\details{
The computations follow page 2319 of Berghaus and Bucher (2018),
  which relates to the estimator given in equation (1.4) on page 2309
  of that paper. Also calculated is an analogous estimate based on the,
  asymptotically equivalent, estimator proposed in Northrop (2015),
  given in equation (1.3) on page 2309 of Berghaus and Bucher (2018).
}
\examples{
spm_sigmahat_dj(newlyn, 20)
}
\references{
Berghaus, B., Bucher, A. (2018) Weak convergence of a pseudo
maximum likelihood estimator for the extremal index. \emph{Ann. Statist.}
\strong{46}(5), 2307-2335. \url{http://dx.doi.org/10.1214/17-AOS1621}

Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{http://dx.doi.org/10.1007/s10687-015-0221-5}
}
\seealso{
\code{\link{spm}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.
}
