% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric_maxima.R
\name{spmax_mle}
\alias{spmax_mle}
\title{Semiparametric maxima estimator of the extremal index}
\usage{
spmax_mle(data, b, sliding = TRUE, F_adjust = TRUE, constrain = TRUE,
  N_or_BB = "N")
}
\arguments{
\item{data}{A numeric vector of raw data.}

\item{b}{A numeric scalar.  The block size.}

\item{sliding}{A logical scalar.}

\item{F_adjust}{A logical scalar.}

\item{constrain}{A logical scalar.}

\item{N_or_BB}{A character scalar.}
}
\value{
Add details.
}
\description{
Add description
}
\details{
Explain re treatment of missings
}
\examples{
spmax_mle(newlyn, 20)
spmax_mle(newlyn, 20, sliding = FALSE)
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
  estimator of the extremal index \emph{Extremes}, \strong{18}(4),
  585-603. \url{http://dx.doi.org/10.1007/s10687-015-0221-5}
}
\seealso{
\code{\link{disjoint_maxima}} for the calculation of the maxima
  over disjoint blocks.

\code{\link{sliding_maxima}} for the calculation of the maxima
  over sliding blocks.

\code{\link{kgaps_mle}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.
}
