% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{summary.dgaps}
\alias{summary.dgaps}
\title{Summary method for a \code{"dgaps"} object}
\usage{
\method{summary}{dgaps}(
  object,
  se_type = c("observed", "expected"),
  digits = max(3, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{object}{an object of class "dgaps", a result of a call to
\code{\link{dgaps}}.}

\item{se_type}{A character scalar. Should the estimate of the standard error
be based on the observed information or the expected information?}

\item{digits}{An integer. Used for number formatting with
\code{\link[base:Round]{signif}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
Returns a list containing the list element \code{object$call}
  and a numeric matrix \code{summary} giving the estimate of the extremal
  index \eqn{\theta} and the estimated standard error (Std. Error).
}
\description{
\code{summary} method for class \code{"dgaps"}
}
\section{Examples}{

See the examples in \code{\link{dgaps}}.
}

\seealso{
\code{\link{dgaps}} for estimation of the extremal index
  \eqn{\theta} using a semiparametric maxima method.

\code{\link{confint.dgaps}} for estimation of confidence intervals
  for \eqn{\theta}.
}
