% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshplot.R
\name{threshplot}
\alias{threshplot}
\title{Plot of Theta against Quantile Level of threshold}
\usage{
threshplot(data, tmin, tmax, conf = 95, k = 1, ...)
}
\arguments{
\item{data}{A numeric vector of raw data.  No missing values are allowed.}

\item{tmin}{A numeric scalar.  Minimum quantile level of threshold applied
to data.}

\item{tmax}{A numeric scalar.  Maximum quantile level of threshold applied
to data.}

\item{conf}{A numeric scalar in (0,100). Confidence level required for the
confidence intervals.}

\item{k}{A numeric scalar.  Run parameter \eqn{K}, as defined in Suveges and
Davison (2010).  Threshold inter-exceedances times that are not larger
than \code{k} units are assigned to the same cluster, resulting in a
\eqn{K}-gap equal to zero.  Specifically, the \eqn{K}-gap \eqn{S}
corresponding to an inter-exceedance time of \eqn{T} is given by
\eqn{S = max(T - K, 0)}.}

\item{...}{Arguments to be passed to methods, such as}
}
\value{
A list containing
  \itemize{
    \item {\code{theta_mle} : } {The maximum likelihood estimate (MLE) of
      \eqn{\theta}.}
    \item {\code{theta_se} : } {The estimated standard error of the MLE.}
    \item {\code{theta_ci} : } {(If \code{conf} is supplied) a numeric
      vector of length two giving lower and upper confidence limits for
      \eqn{\theta}.}
    \item {\code{ss} : } {The list of summary statistics returned from
      \code{\link{kgaps_stats}}.}
  }
}
\description{
Calculates maximum likelihood estimates of the extremal index \eqn{\theta}
based on the K-gaps model for threshold inter-exceedances times of
Suveges and Davison (2010).
}
\details{
The maximum likelihood estimate of the extremal index \eqn{\theta}
  under the K-gaps model of Suveges and Davison (2010) is calculated.
  The form of the log-likelihood is given in the \strong{Details} section of
   \code{\link{kgaps_stats}}. The plot produced presents the movement of the
   extremal index \eqn{\theta} along the quantile level of threshold
   indicated by the user. Confidence intervals are also presented on the
   graph. The confidence level can be chosen by the user, using the argument
   \code{conf}.
}
\examples{
threshplot(newlyn, 40, 80)
}
\references{
Coles, S.G. An Introduction to Statistical Modeling of Extreme
Values (2001)
\url{http://dx.doi.org/10.1007/978-1-4471-3675-0}
}
\seealso{
\code{\link{kgaps_mle}} for maximum likelihood estimation for the
  K-gaps model.
}
