% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex.R
\docType{data}
\name{uccle}
\alias{uccle}
\title{Uccle maximum daily temperatures}
\format{
A data frame with 65036 observations on the following and 5
  variables.
\itemize{
\item{\code{temp:}}{ daily maximum temperature in degrees C.}
\item{\code{year:}}{ the year.}
\item{\code{month:}}{ the month of the year.}
\item{\code{day:}}{ day of the month.}
\item{\code{date:}}{ date with the \code{\link[base:Dates]{Date}} class,
  in the format YYYY-MM-DD.}
}
}
\source{
Klein Tank, A.M.G. and Coauthors, 2002. Daily dataset of
 20th-century surface air temperature and precipitation series for the
 European Climate Assessment. \emph{Int. J. of Climatol.}, \strong{22},
 1441-1453 \doi{10.1002/joc.773}. Data and metadata available at
 \href{https://www.ecad.eu}{https://www.ecad.eu}.  The data were downloaded
 on 27/3/2022 using a
 \href{https://www.ecad.eu/dailydata/customquery.php}{Custom query (ASCII)},
 selecting "non-blend" for type of series.
}
\usage{
uccle
}
\description{
The dataframe \code{uccle} contains daily maximum temperatures in degrees C
recorded at the Uccle, Belgium from 1/1/1833 to 23/1/2011.  The Station
identifier in the source file is 17 and the Source identifier is 117882.
}
\note{
There are 5336 missing values.
}
\keyword{datasets}
