% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaps_methods.R
\name{vcov.dgaps}
\alias{vcov.dgaps}
\title{Calculate Variance-Covariance Matrix for a \code{"dgaps"} object}
\usage{
\method{vcov}{dgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("dgaps", "exdex")} returned from
\code{\link{dgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A 1 by 1 numeric matrix containing the estimated variance of the
  estimator.
}
\description{
\code{vcov} method for class \code{c("dgaps", "exdex")}.
}
