% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_b.R
\name{choose_b}
\alias{choose_b}
\title{Block length diagnostic plot}
\usage{
choose_b(data, b, sliding = TRUE, bias_adjust = c("BB3", "BB1", "N",
  "none"), constrain = TRUE, varN = TRUE, parm = c("both", "N2015",
  "BB2018"), level = 0.95, conf_scale = c("theta", "log"),
  interval_type = c("lik", "sym"), type = c("vertical", "cholesky",
  "spectral", "none"))
}
\description{
Description
}
\examples{
#res <- choose_b(newlyn, 20:60)
#plot(res)
}
