% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgaps_methods.R
\name{coef.kgaps}
\alias{coef.kgaps}
\title{Extract Model Coefficients from a \code{"kgaps"} object}
\usage{
\method{coef}{kgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("kaps", "exdex")} returned from
\code{\link{kgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A numeric scalar: the estimate of the extremal index \eqn{\theta}.
}
\description{
\code{coef} method for class \code{c("kgaps", "exdex")}.
}
