% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confint.exdex}
\alias{confint.exdex}
\title{Confidence intervals for the extremal index \eqn{\theta}}
\usage{
\method{confint}{exdex}(object, parm = c("both", "N2015", "BB2018"),
  maxima = c("sliding", "disjoint"), level = 0.95, constrain = TRUE,
  conf_scale = c("theta", "log_theta"), bias_adjust = TRUE,
  type = c("vertical", "cholesky", "spectral", "none"), plot = FALSE,
  ndec = 2, ...)
}
\arguments{
\item{object}{An object of class \code{"exdex"}, returned by
\code{\link{spm}}.}

\item{parm}{A character scalar specifying whether to estimate confidence
intervals for both variants of the estimator (\code{"both"}),
for the Northrop (2015) variant only (\code{"N2015"}) or
for the Berhaus and Bucher (2018) variant only (\code{"BB2018"}).}

\item{maxima}{A character scalar specifying whether to estimate
confidence intervals based on sliding maxima or disjoint maxima.}

\item{level}{The confidence level required.  A numeric scalar in (0, 1).}

\item{constrain}{A logical scalar.  If \code{constrain = TRUE} then
any confidence limits that are greater than 1 are set to 1,
that is, they are constrained to lie in (0, 1].  Otherwise,
limits that are greater than 1 may be obtained.
If \code{constrain = TRUE} then any lower confidence limits that are
less than 0 are set to 0.}

\item{conf_scale}{A character scalar.  Determines the scale on which
  we use approximate large-sample normality of the estimators to
  estimate confidence intervals.

  If \code{conf_scale = "theta"}
  then confidence intervals are estimated for \eqn{\theta} directly.
  If \code{conf_scale = "log_theta"} then confidence intervals are first
  estimated for \eqn{\log\theta}{log\theta} and then transformed back
  to the \eqn{\theta}-scale.

  Any bias-adjustment requested in the original call to \code{\link{spm}},
  using it's \code{bias_adjust} argument, is automatically applied here.}

\item{bias_adjust}{A logical scalar.  If \code{bias_adjust = TRUE} then,
  if appropriate, bias-adjustment is also applied to the loglikelihood
  before it is adjusted using \code{\link[chandwich]{adjust_loglik}}.
  This is performed only if, in the call to
  \code{\link{spm}}, \code{bias_adjust = "BB3"} or
  \code{"BB1"} was specified, that is, we have
  \code{object$bias_adjust = "BB3"}
  or \code{"BB1"}.  In these cases the relevant
  component of \code{object$bias_val} is used to scale \eqn{\theta} so
  that the location of the maximum of the loglikelihood lies at the
  bias-adjusted estimate of \eqn{\theta}.

  If \code{bias_adjust = FALSE} or \code{object$bias_adjust = "none"}
  or \code{"N"} then no bias-adjustment of the
  intervals is performed.  In the latter case this is because the
  bias-adjustment is applied in the creation of the data in
  \code{object$N2015_data} and \code{object$BB2018_data}, on which the
  naive likelihood is based.}

\item{type}{A character scalar.  The argument \code{type} to be passed to
\code{\link[chandwich]{conf_intervals}} in order to estimate the
likelihood-based intervals.  See \strong{Details}.
Using \code{type = "none"} is \emph{not} advised because then the
intervals are based on naive estimated standard errors.  In particular,
if (the default) \code{sliding = TRUE} was used in the call to
\code{\link{spm}} then the likelihood-based confidence intervals provide
\emph{vast} underestimates of uncertainty.}

\item{plot}{A logical scalar.  If \code{plot = TRUE} then a plot of the
adjust loglikelihood is produced, using
\code{\link[chandwich]{plot.confint}}.}

\item{ndec}{An integer scalar.  The legend (if included on the plot)
contains the confidence limits rounded to \code{ndec} decimal places.}

\item{...}{Further arguments to be passed to
\code{\link[chandwich]{plot.confint}}, used only if
\code{plot = TRUE}.}
}
\value{
A matrix with columns giving the lower and upper confidence limits.
  These will be labelled as (1 - level)/2 and 1 - (1 - level)/2 in \%
  (by default 2.5\% and 97.5\%).
  The row names are a concatentation of the variant of the estimator
  (N2015 for Northrop (2015), BB2018 for Berghaus and Bucher (2018))
  and the type of interval
  (sym for symmetric and lik for likelihood-based).
}
\description{
\code{confint} method for objects of class \code{"exdex"}.
Computes confidence intervals for \eqn{\theta} based on an object returned
from \code{\link{spm}}.  Two types of interval are returned:
(a) intervals based on approximate large-sample normality of the estimators
of \eqn{\theta}, which are symmetric about the respective point estimates,
and (b) likelihood-based intervals based on an adjustment of a naive
(pseudo-) loglikelihood, using the \code{\link[chandwich]{adjust_loglik}}
function in the \code{\link[chandwich]{chandwich}} package.
}
\details{
The likelihood-based intervals are estimated using the
  \code{\link[chandwich]{adjust_loglik}} function in the
  \code{\link[chandwich]{chandwich}} package, followed by a call to
  \code{\link[chandwich]{conf_intervals}}.

  If \code{object$se} contains \code{NA}s, because the block size \code{b}
  was too small or too large in the call to \code{\link{spm}} then
  confidence intervals cannot be estimated and an error will be thrown.
  See the \strong{Details} section of the \code{\link{spm}} documentation
  for more information.
}
\examples{
res <- spm(newlyn, 20)
# I can't include these examples until new chandwich is on CRAN
#confint(res)
#confint(res, plot = TRUE)
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{https://doi.org/10.1007/s10687-015-0221-5}

Berghaus, B., Bucher, A. (2018) Weak convergence of a pseudo
maximum likelihood estimator for the extremal index. \emph{Ann. Statist.}
\strong{46}(5), 2307-2335. \url{https://doi.org/10.1214/17-AOS1621}
}
