% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex-internal.R
\name{exdex-internal}
\alias{exdex-internal}
\alias{spm_check}
\alias{spm_sigmahat_dj}
\alias{spm_R_quick}
\alias{ests_sigmahat_dj}
\alias{log0const_slow}
\alias{log0const}
\alias{ecdf3}
\alias{ecdf2}
\alias{ecdf1}
\title{Internal exdex functions}
\usage{
spm_check(data, b, sliding = TRUE, bias_adjust = c("BB3", "BB1", "N",
  "none"), constrain = TRUE, varN = TRUE, which_dj = c("last",
  "first"))

spm_sigmahat_dj(data, b, check = FALSE)

spm_R_quick(data, b, bias_adjust = c("BB3", "BB1", "N", "none"),
  constrain = TRUE, varN = TRUE, which_dj = c("last", "first"))

ests_sigmahat_dj(all_max, b, which_dj, bias_adjust)

log0const_slow(x, const)

log0const(x, const)

ecdf3(x, y)

ecdf2(x, y)

ecdf1(x, y, lenx)
}
\description{
Internal exdex functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
