% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex.R
\docType{package}
\name{exdex}
\alias{exdex}
\alias{exdex-package}
\title{exdex: Estimation of the Extremal Index}
\description{
Performs frequentist inference for the extremal index using the methodologies
described in Suveges and Davison (2010)  \url{http://dx.doi.org/10.1214/09-AOAS292} and in
Northrop (2015) \url{http://dx.doi.org/10.1007/s10687-015-0221-5}.
This is done in two ways, based on the two main forms of extreme value theory.
One uses threshold inter-exceedance times based on the K-gaps model and the
other one is based on block maxima.
}
\details{
The main functions in the exdex package are \code{\link{kgaps_mle}}
and \code{\link{spm_mle}} which are used to estimate the extremal index
\eqn{\theta}. The user chooses the extreme value model and provides the data.
Additionally, the user for the \eqn{K}-gaps model chooses an extreme value
threshold \eqn{thresh} that is applied to data, whereas for the block maxima
method he chooses \eqn{b}, the block size.

  See \code{vignette("exdex-vignette", package = "exdex")} for an
  overview of the package.
}
\references{
Suveges, M. and Davison, A. C. (2010) Model
  misspecification in peaks over threshold analysis, \emph{The Annals of
  Applied Statistics}, \strong{4}(1), 203-221.
  \url{http://dx.doi.org/10.1214/09-AOAS292}

Attalides, N. (2015) Threshold-based extreme value modelling,
  PhD thesis, University College London.
  \url{http://discovery.ucl.ac.uk/1471121/1/Nicolas_Attalides_Thesis.pdf}
}
\seealso{
\code{\link{kgaps_mle}} to calculate the maximum likelihood estimates
of the extremal index \eqn{\theta} based on the K-gaps model for threshold
inter-exceedances times of Suveges and Davison (2010).

\code{\link{spm_mle}} to calculate the semiparametric maxima
estimator of the extremal index \eqn{\theta} based on sliding or disjoint
block maxima based on Northrop (2015).
}
