% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdex.R
\docType{package}
\name{exdex}
\alias{exdex}
\alias{exdex-package}
\title{exdex: Estimation of the Extremal Index}
\description{
Performs frequentist inference for the extremal index using the methodologies
described in Suveges and Davison (2010)  \url{http://dx.doi.org/10.1214/09-AOAS292} and in
Northrop (2015) \url{http://dx.doi.org/10.1007/s10687-015-0221-5}.
}
\details{
Add details.

  See \code{vignette("exdex-vignette", package = "exdex")} for an
  overview of the package.
}
\references{
Suveges, M. and Davison, A. C. (2010) Model
  misspecification in peaks over threshold analysis, \emph{The Annals of
  Applied Statistics}, \strong{4}(1), 203-221.
  \url{http://dx.doi.org/10.1214/09-AOAS292}

Attalides, N. (2015) Threshold-based extreme value modelling,
  PhD thesis, University College London.
  \url{http://discovery.ucl.ac.uk/1471121/1/Nicolas_Attalides_Thesis.pdf}
}
