% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{plot.confint_spm}
\alias{plot.confint_spm}
\title{Plot diagnostics for a confint_spm object}
\usage{
\method{plot}{confint_spm}(x, y = NULL, ndec = 2, ...)
}
\arguments{
\item{x}{an object of class c("confint_spm", "exdex"), a result of a call
to \code{\link{confint.spm}}.}

\item{y}{Not used.}

\item{ndec}{An integer scalar.  The legend (if included on the plot)
contains the confidence limits rounded to \code{ndec} decimal places.}

\item{...}{Further arguments to be passed to
\code{\link[chandwich]{plot.confint}}.}

\item{legend}{A logical scalar or a character vector.  If this is
supplied then a legend is added to the plot.  If \code{legend} is a
character vector then it is used as the argument \code{legend}
to \code{\link[graphics]{legend}}.  Otherwise, i.e. if
\code{legend = TRUE} then the argument \code{type} is used.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an objects of class c("confint_spm", "exdex").
}
\examples{
res <- spm(newlyn, 20)
# I can't include these examples until new chandwich is on CRAN
#cis <- confint(res)
# plot(cis)
}
