% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_maxima.R
\name{spm}
\alias{spm}
\title{Semiparametric maxima estimator of the extremal index}
\usage{
spm(data, b, sliding = TRUE, bias_adjust = c("BB3", "BB1", "N",
  "none"), constrain = TRUE, varN = TRUE)
}
\arguments{
\item{data}{A numeric vector of raw data.}

\item{b}{A numeric scalar.  The block size.}

\item{sliding}{A logical scalar.  If \code{sliding = TRUE} then the
estimates are based on sliding block maxima.  Otherwise, the estimates
are based on disjoint block maxima.}

\item{bias_adjust}{A character scalar.  Is bias-adjustment of the
raw estimate of \eqn{\theta} performed using the bias-reduced
estimator (\code{bias_adjust = "BB3"}), derived in Section 5 of
Berghaus and Bucher (2018); or a simpler version of this
(\code{bias_adjust = "BB1"}), in which the raw estimate is multiplied by
\eqn{(k-1) / k}, where \eqn{k} is the number of blocks; or the
bias-adjustment of the empirical distribution function used to calculate
the estimate, as detailed in Section 2 of Northrop (2015).  When disjoint
maxima are used \code{bias_adjust = "BB1"} and \code{bias_adjust = "N"}
give identical estimates of the Berghaus and Bucher (2018) variant,
as explained at the end of Section 5 of Berghaus and Bucher (2018).
If \code{bias_adjust = "none"} then no bias-adjustment is performed.}

\item{constrain}{A logical scalar.  If \code{constrain = TRUE} then
any estimates that are greater than 1 are set to 1,
that is, they are constrained to lie in (0, 1].  This is carried out
\emph{after} any bias-adjustment.  Otherwise,
estimates that are greater than 1 may be obtained.}

\item{varN}{A logical scalar.  If \code{varN = TRUE} then the estimation
of the sampling variance of the Northrop (2015) estimator is tailored
to that estimator.  Otherwise, the sampling variance derived in
Berghaus and Bucher (2018) is used.
See \strong{Details} for further information.}
}
\value{
A list containing
  \itemize{
    \item {\code{theta} : } {A vector containing the estimates \eqn{\theta}
      resulting from the two variants of the semiparametric estimator,
      labelled N2015 for Northrop (2015) and BB2018 for
      Berghaus and Bucher (2018).}
    \item {\code{se} : } {The respective estimated standard errors.}
    \item {\code{uncontrained_theta} : } {The estimates of \eqn{\theta}
      without the constraint that they lie in (0, 1])}
    \item {\code{N2015_data} : } {The values of the \eqn{Y}-data.}
    \item {\code{BB2018_data} : } {The values of the \eqn{Z}-data.}
    \item {\code{bias_val} : } {The respective values of the
      bias-adjustment applied to the raw estimates.  This is only relevant
      if \code{bias_adjust} is "BB3" or "BB1".  Otherwise, \code{bias_val}
      is \code{NA}.}
    \item {\code{bias_adjust} : } {The input value of \code{bias_adjust}.}
    \item {\code{b} : } {The input value of \code{b}.}
    \item {\code{sliding} : } {The input value of \code{sliding}.}
    \item {\code{call} : } {The call to \code{spm}.}
  }
}
\description{
Estimates the extremal index \eqn{\theta} using a semiparametric block
maxima estimator of Northrop (2015) and a variant of this estimator
studied by Berghaus and Bucher (2018).  Estimates of uncertainty are
calculated using the asymptotic theory developed by Berghaus and
Bucher (2018).
}
\details{
The extremal index \eqn{\theta} is estimated using the
  semiparametric maxima estimator of Northrop (2015) and variant
  of this studied by Berghaus and Bucher (2018).  In each case a sample
  of 'data' is derived from the input data \code{data}, based on the
  empirical distribution function of these data evaluated at the
  maximum values of  of blocks of \code{b} contiguous values in \code{data}.

  The estimators are based on an assumption that these 'data' are sampled
  approximately from an exponential distribution with mean \eqn{1/\theta}.
  For details see page 2309 of Berghaus and Bucher (2018), where the
  'data' for the Northrop (2015) estimator are denoted \eqn{Y} and
  those for the Berghaus and Bucher (2018) are denoted \eqn{Z}.
  For convenience, we will refer to these values as the
  \eqn{Y}-data and the \eqn{Z}-data.

  If \code{sliding = TRUE} then the function uses sliding block maxima,
  that is, the largest value observed in \emph{all}
  (\code{length(data) - b + 1}) blocks of \code{b} observations.
  If \code{sliding = FALSE} then disjoint block maxima, that is, the
  largest values in (\code{floor(n / b)}) disjoint blocks of \code{b}
  observations, are used.

  Estimation of the sampling variances of the estimators is based
  on Proposition 4.1 of Berghaus and Bucher (2018).
  For the Northrop (2015) variant the user has the choice either to
  use the sampling variance based on the Berghaus and Bucher (2018)
  estimator, i.e. the \eqn{Z}-data (\code{varN = FALSE}) or an analogous
  version tailored to the Northrop (2015) estimator that uses the
  \eqn{Z}-data (\code{varN = TRUE}).

  A condition imposed in Proposition 4.1 of Berghaus and Bucher (2018)
  means that \code{b} must be no smaller than \eqn{k^{1/2}} and no larger
  than \eqn{k^2}, where \eqn{k} is \code{floor(length(data) / b)},
  i.e. \eqn{k} is the number of complete blocks.  If this is not the case
  then a warning will be given and
    \itemize{
      \item{estimated standard errors will be missing from the returned
            object,}
      \item{if \code{bias_adjust == "BB3"} then this bias-adjustment
            based on \code{bias_adjust == "BB1"} will be performed instead,
            because the former relies on the estimated variances of the
            estimators.}
    }
}
\examples{
temp <- spm(-as.vector(sp500[2:6550]), 250)

temp <- spm(newlyn, 20)
temp <- spm(newlyn, 20, sliding = FALSE)
}
\references{
Northrop, P. J. (2015) An efficient semiparametric maxima
estimator of the extremal index. \emph{Extremes} \strong{18}(4), 585-603.
\url{https://doi.org/10.1007/s10687-015-0221-5}

Berghaus, B., Bucher, A. (2018) Weak convergence of a pseudo
maximum likelihood estimator for the extremal index. \emph{Ann. Statist.}
\strong{46}(5), 2307-2335. \url{https://doi.org/10.1214/17-AOS1621}
}
\seealso{
\code{\link{kgaps_mle}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.

\code{\link{confint.exdex}} to estimate confidence intervals
  for \eqn{theta}.
}
