% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_maxima.R
\name{spm_mle}
\alias{spm_mle}
\title{Semiparametric maxima estimator of the extremal index}
\usage{
spm_mle(data, b, sliding = TRUE, constrain = TRUE, conf = NULL,
  R = 1000)
}
\arguments{
\item{data}{A numeric vector of raw data.}

\item{b}{A numeric scalar.  The block size.}

\item{sliding}{A logical scalar indicating whether use sliding blocks
(\code{TRUE}) or disjoint blocks (\code{FALSE}).}

\item{constrain}{A logical scalar indicating whether or not to constrain the
mle to lie in the interval (0, 1].}

\item{conf}{A numeric scalar.  If \code{conf} is supplied then the
the standard error and \code{conf}\% likelihood-based confidence interval
for \eqn{\theta}, and the standard error of the estimator of \eqn{\theta},
are estimated using block bootstrapping, implemented by
\code{\link[boot]{tsboot}}.  See Northrop (2015) and references therein
for more information.}

\item{R}{An integer scalar.  The number of bootstrap resamples used to
estimate the standard error and confidence intervals.
See \code{\link[boot]{tsboot}}.}
}
\value{
A list containing
  \itemize{
    \item {\code{theta_mle} : } {The maximum likelihood estimate (MLE) of
      \eqn{\theta}.}
    \item {\code{theta_se} : } {The estimated standard error of the MLE.}
    \item {\code{theta_ci} : } {(If \code{conf} is supplied) a numeric
      vector of length two giving lower and upper confidence limits for
      \eqn{\theta}.}
  }
  If \code{conf} is not supplied then only the MLE \code{theta_mle}
  is returned.
}
\description{
Calculates the semiparametric maxima estimator of the extremal index
\eqn{\theta} based on sliding or disjoint block maxima based on Northrop (2015).
}
\details{
The extremal index \eqn{\theta} is estimated using the semiparametric
maxima estimator of Northrop (2015).  If \code{sliding = TRUE} then the
function uses sliding block maxima, that is, the largest value observed in
\emph{all} blocks of \code{b} observations, whereas if \code{sliding = FALSE}
then disjoint block maxima, that is, the largest values in non-overlapping
blocks of \code{b} observations, are used.  If \code{constrain = TRUE} then
if the raw estimate of the extremal index is greater than one then a value of
1 is returned. Otherwise (\code{constrain = FALSE}) the raw estimate is
returned, even if it is greater than 1.
}
\examples{
spm_mle(newlyn, 20)
spm_mle(newlyn, 20, sliding = FALSE)

spm_mle(newlyn, 20, sliding = FALSE, conf = 95)
# When sliding = TRUE bootstrapping is slow
\dontrun{
spm_mle(newlyn, 20, conf = 95)
}
}
\references{
Northrop, P. J. (2015) \emph{An efficient semiparametric maxima
estimator of the extremal index} Extremes, \strong{18}(4), 585-603.
\url{http://dx.doi.org/10.1007/s10687-015-0221-5}
}
\seealso{
\code{\link{kgaps_mle}} for maximum likelihood estimation of the
  extremal index \eqn{\theta} using the K-gaps model.
}
