% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgaps_methods.R
\name{vcov.kgaps}
\alias{vcov.kgaps}
\title{Calculate Variance-Covariance Matrix for a \code{"kgaps"} object}
\usage{
\method{vcov}{kgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("kgaps", "exdex")} returned from
\code{\link{kgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A 1 by 1 numeric matrix containing the estimated variance of the
  estimator.
}
\description{
\code{vcov} method for class \code{c("kgaps", "exdex")}.
}
