% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgaps_methods.R
\name{vcov.kgaps}
\alias{vcov.kgaps}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{kgaps}(object, ...)
}
\arguments{
\item{object}{and object of class \code{c("kgaps", "exdex")} returned from
\code{\link{kgaps}}.}

\item{...}{Further arguments.  None are used.}
}
\value{
A 1 by 1 numeric matrix containing the estimated variance of the
  estimator.
}
\description{
\code{vcov} method for class \code{c("kgaps", "exdex")}.
}
\references{
Suveges, M. and Davison, A. C. (2010) Model
  misspecification in peaks over threshold analysis, \emph{The Annals of
  Applied Statistics}, \strong{4}(1), 203-221.
  \url{https://doi.org/10.1214/09-AOAS292}

Attalides, N. (2015) Threshold-based extreme value modelling,
  PhD thesis, University College London.
  \url{http://discovery.ucl.ac.uk/1471121/1/Nicolas_Attalides_Thesis.pdf}
}
